# src/main/java/com/github/mattthey/api/dto/AuthorDto.java
package com.github.mattthey.api.dto;

import java.time.OffsetDateTime;
import java.util.List;

public record AuthorDto(
        Long id,
        String name,
        OffsetDateTime dateOfBirth,
        OffsetDateTime dateOfDeath
) {
}